requirements.txt
pytorch-ignite
pyyaml
tqdm
mlflow
sagemaker
torch
torchvision
aws-sagemaker-remote
